+++ assets/js/app.js
@@ -26,6 +26,8 @@ import 'phoenix_html'
 import { Socket } from 'phoenix'
 import { LiveSocket } from 'phoenix_live_view'
 import topbar from 'topbar'
+
+import CtrlEnter from './ctrlenter'
 import Date from './date'
 import DateTime from './datetime'
 
@@ -34,7 +36,7 @@ let csrfToken
 if (csrfTokenElement) { csrfToken = csrfTokenElement.getAttribute('content') }
 const liveSocket = new LiveSocket('/live', Socket, {
   params: { _csrf_token: csrfToken },
-  hooks: { Date, DateTime }
+  hooks: { CtrlEnter, Date, DateTime }
 })
 
 // Show progress bar on live navigation and form submits