+++ changelog.md
@@ -7,6 +7,7 @@
 - Make content previews resizable
 - Fix live flashes not dismissable by click
 - Fix disconnection modal not displaying
+- Submit items with ctrl-enter
 
 # v0.1.12
 - Code quality fixes